A good first draft to automatically detect MediaWiki config parameters, in Git history and some heuristics to detect type
This script explores the whole Git history of MediaWiki for the file includes/DefaultSettings.php. It creates a list of
config parameters added at some point in history, and try to detect and follow type of the parameter as well as
default value of the parameter. The raw picture works well.

Some types are still not correctly understood (numbers, functions), and it should be decided how to
manage edge cases where parameter values comes from a function (how to write it down?).