The schema for config parameters can be created almost without schema errors.
Added an early detection of the type to better compute:
* transitive types ($wgMyVar = &$wgMyOtherVar)
* result of number computations
* transitive results

All remaining types have been added (objects, numbers, integers).

Errors were found in UTF-8 encoding of DefaultSettings.php; specific commits were added to
replace faulty characters. Without it, json_encode outputs nothing (and JSON standard says
Unicode SHALL be used).

Tested on the whole history of DefaultSettings.php from initial commit in 2003 to
MediaWiki 1.27.0-alpha in 2016. Works reasonnably correctly. 63 schema violations
remains, but these are of two types:
* use of constants (mainly cache config parameters): difficult to retrieve their value
* use of functions (mainly optional features): should be eval’ed but not that great
Given the small number of violations (it decreased from 5000 to 63), the difficulty for a
program, and the facility for a human, this is left as an exercice to a human.